+++ src/components/v2/appDetails/k8Resource/nodeDetail/nodeDetail.api.ts
@@ -92,15 +92,22 @@ function createBody(appDetails: AppDetails, nodeName: string, nodeType: string, 
     const selectedResource = appDetails.resourceTree.nodes.filter(
         (data) => data.name === nodeName && data.kind.toLowerCase() === nodeType,
     )[0]
+
+    const getAppName = (): string => {
+        if (appDetails.deploymentAppType === DeploymentAppType.helm && appDetails.appType === AppType.DEVTRON_APP) {
+            return `${appDetails.appName}-${appDetails.environmentName}`
+        } else {
+           return appDetails.appName
+        }
+    }
+
     const appId =
         appDetails.deploymentAppType == DeploymentAppType.argo_cd
             ? ''
             : getAppId(
                   appDetails.clusterId,
                   appDetails.namespace,
-                  appDetails.deploymentAppType === DeploymentAppType.helm && appDetails.appType === AppType.DEVTRON_APP
-                      ? `${appDetails.appName}-${appDetails.environmentName}`
-                      : appDetails.appName,
+                  getAppName()
        )
 
     const requestBody = {
